Properly determine the input device, when the operator backend is `mixed` (#174)
`DaliPipeline::GetInputDevice` function checks, what is the backend of given operator and returns information, what kind of device (CPU or GPU) shall the input data belong to. It worked well for input device CPU or GPU. It broke when introducing `VideoInput` operator, since the device of this operator can be `mixed`. This PR solves the problem and adds proper test for such situation.

* Fix input device check
* Add tests

---------

Signed-off-by: Rafal <rbanas@nvidia.com>